+++ packages/expo-ui/src/swift-ui/index.tsx
@@ -1,5 +1,3 @@
-import '../State/index.fx';
-
 export * from './AccessoryWidgetBackground';
 export * from './Alert';
 export * from './BottomSheet';